Revert the image handling to what it was previously
I think that the initial configuration is better. This way, the first image of the index is the main one, and the subsequent ones can still be added and used later (say on click you can scroll through them). These images make sense paired, the first image is the "first one showed". It does not make sense to have two different objects for photos.